requirements.txt
docopt>=0.6.2
ffmpeg-python>=0.1.16
termcolor>=1.1.0